requirements.txt
Django==1.9.1
django-tagulous==0.11.1
